+++ Tests/TuistKitTests/Generator/LinkGeneratorTests.swift
@@ -36,7 +36,7 @@ final class LinkGeneratorErrorTests: XCTestCase {
         fileElements.products["waka.framework"] = wakaFile
         let sourceRootPath = AbsolutePath("/")
 
-        try subject.generateEmbedPhase(dependencies: dependencies,
+        try subject.generateEmbedPhase(dependencies: Set(dependencies),
                                        pbxTarget: pbxTarget,
                                        pbxproj: pbxproj,
                                        fileElements: fileElements,
@@ -62,7 +62,7 @@ final class LinkGeneratorErrorTests: XCTestCase {
         let fileElements = ProjectFileElements()
         let sourceRootPath = AbsolutePath("/")
 
-        XCTAssertThrowsError(try subject.generateEmbedPhase(dependencies: dependencies,
+        XCTAssertThrowsError(try subject.generateEmbedPhase(dependencies: Set(dependencies),
                                                             pbxTarget: pbxTarget,
                                                             pbxproj: pbxproj,
                                                             fileElements: fileElements,